ref(core): Split report schemas into module tree
Move core schemas and helpers into focused harness and report modules while keeping the root public exports stable. Tighten owned artifact schemas so provider-specific usage fields stay under metadata.

Fix trace transcript assembly to preserve repeated turns, and make the report UI server resolve packaged assets through an ESM runtime shim instead of stack parsing.
