+++ packages/core/src/report/index.ts
@@ -1,0 +1,45 @@
+export {
+  EvalMetaSchema,
+  EvalScoreSchema,
+  EvalTaskMetaSchema,
+  HarnessMetaSchema,
+  readEvalTaskMeta,
+} from "./metadata";
+export type {
+  EvalMeta,
+  EvalScore,
+  EvalTaskMeta,
+  HarnessMeta,
+} from "./metadata";
+
+export {
+  parseVitestJsonReport,
+  VitestJsonAssertionSchema,
+  VitestJsonFileSchema,
+  VitestJsonLocationSchema,
+  VitestJsonReportSchema,
+  VitestJsonStatusSchema,
+} from "./vitest-json";
+export type {
+  VitestJsonAssertion,
+  VitestJsonFile,
+  VitestJsonLocation,
+  VitestJsonReport,
+  VitestJsonStatus,
+} from "./vitest-json";
+
+export {
+  collectReportWorkspace,
+  parseReportWorkspace,
+  REPORT_WORKSPACE_SCHEMA_VERSION,
+  ReportCaseSchema,
+  ReportRunSchema,
+  ReportWorkspaceSchema,
+} from "./workspace";
+export type {
+  CollectReportWorkspaceOptions,
+  ReportCase,
+  ReportRun,
+  ReportWorkspace,
+  ReportWorkspaceInput,
+} from "./workspace";
