+++ packages/report-ui/src/esm-runtime.d.ts
@@ -1,0 +1,1 @@
+export function currentModuleUrl(): string;
