+++ packages/report-ui/src/esm-runtime.js
@@ -1,0 +1,3 @@
+export function currentModuleUrl() {
+  return import.meta.url;
+}
